# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BR Tours and Travels API
  version: 0.3.0
  description: >-
    Express business API for the public website, authenticated administration,
    media, enquiries and notification operations. Booking requests are leads,
    not payments, issued inventory or guaranteed reservations.
servers:
  - url: /api/v1
...
paths:
...
  /inquiries:
    post:
      tags: [Enquiries]
      summary: Persist an idempotent contact, package enquiry or booking request
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InquiryInput" }
      responses:
        "201": { description: New request persisted with its outbox event }
        "200":
          {
            description: Identical idempotent retry returned the original receipt,
          }
        "409": { $ref: "#/components/responses/Error" }
...
  /admin/inquiries:
    get:
      tags: [Enquiries]
      security: [{ cookieAuth: [] }]
      parameters:
        - { in: query, name: q, schema: { type: string, maxLength: 120 } }
        - {
            in: query,
            name: status,
            schema:
              {
                type: string,
                enum: [NEW, CONTACTED, QUOTED, CONFIRMED, CLOSED, LOST],
              },
          }
        - {
            in: query,
            name: type,
            schema:
              {
                type: string,
                enum: [CONTACT, PACKAGE_ENQUIRY, BOOKING_REQUEST],
              },
          }
        - { in: query, name: package, schema: { type: string, maxLength: 180 } }
        - { in: query, name: from, schema: { type: string, format: date } }
        - { in: query, name: to, schema: { type: string, format: date } }
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PageSize" }
      responses: { "200": { description: Sales-authorized bounded lead list } }
  /admin/inquiries/export.csv:
    get:
      tags: [Enquiries]
      security: [{ cookieAuth: [] }]
      responses: { "200": { description: Formula-injection-safe CSV export } }
  /admin/inquiries/{id}:
    get:
      tags: [Enquiries]
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        {
          "200":
            {
              description: "Private lead detail, notes, history and redacted delivery state",
            },
        }
  /admin/inquiries/{id}/status:
    patch:
      tags: [Enquiries]
      security: [{ cookieAuth: [] }]
      parameters:
        [
          { $ref: "#/components/parameters/Id" },
          { $ref: "#/components/parameters/CsrfToken" },
        ]
      responses: { "200": { description: Audited valid status transition } }
  /admin/inquiries/{id}/notes:
    post:
      tags: [Enquiries]
      security: [{ cookieAuth: [] }]
      parameters:
        [
          { $ref: "#/components/parameters/Id" },
          { $ref: "#/components/parameters/CsrfToken" },
        ]
      responses: { "201": { description: Private staff note created } }
  /admin/inquiries/{id}/assignment:
    patch:
      tags: [Enquiries]
      security: [{ cookieAuth: [] }]
      parameters:
        [
          { $ref: "#/components/parameters/Id" },
          { $ref: "#/components/parameters/CsrfToken" },
        ]
      responses: { "200": { description: Lead assignment updated } }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: br_admin_session
      description: Environment-configured opaque session cookie; never a browser-storage token.
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string, maxLength: 30 }
...
    Page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, maximum: 1000, default: 1 }
    PageSize:
      in: query
      name: pageSize
      schema: { type: integer, minimum: 1, maximum: 48, default: 12 }
    CsrfToken:
      in: header
      name: x-csrf-token
      required: true
      schema: { type: string }
    IdempotencyKey:
      in: header
      name: idempotency-key
      required: true
      schema: { type: string, minLength: 16, maxLength: 128 }
  schemas:
...
    InquiryInput:
      type: object
      additionalProperties: false
      required: [type, name, email, message, privacyAccepted, policyVersion]
      properties:
        type:
          { type: string, enum: [CONTACT, PACKAGE_ENQUIRY, BOOKING_REQUEST] }
        name: { type: string, minLength: 2, maxLength: 120 }
        email: { type: string, format: email, maxLength: 254 }
        phone: { type: string, maxLength: 40 }
        subject: { type: string, maxLength: 200 }
        message: { type: string, minLength: 10, maxLength: 5000 }
        packageSlug: { type: string, maxLength: 180 }
        departureId: { type: string, maxLength: 30 }
        preferredStartDate: { type: string, format: date }
        adultCount: { type: integer, minimum: 1, maximum: 50 }
        childCount: { type: integer, minimum: 0, maximum: 50 }
        budget: { type: number, minimum: 0 }
        sourcePath: { type: string, maxLength: 500 }
        privacyAccepted: { type: boolean, const: true }
        policyVersion: { type: string, maxLength: 64 }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            fields:
              {
                type: object,
                additionalProperties: { type: array, items: { type: string } },
              }
            requestId: { type: string }
  responses:
    Error:
      description: Safe API error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }